requirements.txt
streamlit>=1.30
scikit-learn>=1.3
pandas>=2.0
numpy>=1.24
joblib>=1.3
requests>=2.31
beautifulsoup4>=4.12
pytesseract>=0.3.10
Pillow>=10.0
transformers>=4.35
torch>=2.0
